pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "enterprise-rag-chatbot"
version = "1.0.0"
description = "AI-powered knowledge management with OCR support for Thai, English, and Japanese documents"
readme = "README.md"
authors = [
    {name = "Enterprise RAG Team"},
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.10"
dependencies = [
    "pydantic-ai>=0.0.14",
    "openai>=1.0.0",
    "chromadb>=0.4.0",
    "easyocr>=1.7.0",
    "pytesseract>=0.3.10",
    "Pillow>=10.0.0",
    "opencv-python>=4.8.0",
    "fastapi>=0.104.0",
    "streamlit>=1.28.0",
    "uvicorn[standard]>=0.24.0",
    "pypdf>=3.17.0",
    "python-docx>=0.8.11",
    "python-multipart>=0.0.6",
    "pandas>=2.1.0",
    "numpy>=1.24.0",
    "aiofiles>=23.0.0",
    "structlog>=23.1.0",
    "loguru>=0.7.0",
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.0.0",
    "langdetect>=1.0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/your-org/enterprise-rag-chatbot"
Documentation = "https://github.com/your-org/enterprise-rag-chatbot/docs"
Repository = "https://github.com/your-org/enterprise-rag-chatbot"
Issues = "https://github.com/your-org/enterprise-rag-chatbot/issues"

[project.scripts]
rag-server = "src.api.main:main"
rag-web = "src.web.streamlit_app:main"